And yet one more MacOS ambiguous cast fix

+++ src/ExportGLTF.cpp
@@ -530,7 +530,7 @@ namespace {
       rjPrimitive.AddMember("mode", 0x0004 /* GL_TRIANGLES */, alloc);
       rjPrimitive.AddMember("attributes", rjAttributes, alloc);
       rjPrimitive.AddMember("indices", indicesAccesorIx, alloc);
-      rjPrimitive.AddMember("material", geos[0].sortKey >> 1, alloc);
+      rjPrimitive.AddMember("material", static_cast<uint32_t>(geos[0].sortKey >> 1), alloc);
 
       rjPrimitives.PushBack(rjPrimitive, alloc);
